refactor: centralize fe mocking utils
