+++ operate/client/package.json
@@ -63,6 +63,7 @@
   },
   "devDependencies": {
     "@camunda/lint-config": "file:../../webapp/client/packages/lint-config",
+    "@camunda/mock-server-utils": "file:../../webapp/client/packages/mock-server-utils",
     "@eslint/js": "9.39.4",
     "@playwright/test": "1.58.2",
     "@testing-library/dom": "10.4.1",
