+++ tasklist/client/package.json
@@ -61,6 +61,7 @@
   "devDependencies": {
     "@axe-core/playwright": "4.11.1",
     "@camunda/lint-config": "file:../../webapp/client/packages/lint-config",
+    "@camunda/mock-server-utils": "file:../../webapp/client/packages/mock-server-utils",
     "@eslint/js": "9.39.4",
     "@playwright/test": "1.58.2",
     "@tanstack/react-query-devtools": "5.91.3",
