+++ tasklist/client/src/modules/mocks/license.tsx
@@ -6,48 +6,10 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {addDays} from 'date-fns';
-import type {License} from '@camunda/camunda-api-zod-schemas/8.10';
-
-const validLicense: License = {
-  validLicense: true,
-  licenseType: 'production',
-  isCommercial: false,
-  expiresAt: new Date().toISOString(),
-};
-
-const saasLicense: License = {
-  validLicense: true,
-  licenseType: 'saas',
-  isCommercial: false,
-  expiresAt: new Date().toISOString(),
-};
-
-const invalidLicense: License = {
-  validLicense: false,
-  licenseType: 'production',
-  isCommercial: false,
-  expiresAt: null,
-};
-
-const commercialExpired: License = {
-  validLicense: true,
-  licenseType: 'production',
-  isCommercial: true,
-  expiresAt: new Date().toISOString(),
-};
-
-const validNonCommercial = {
-  validLicense: true,
-  licenseType: 'production',
-  isCommercial: false,
-  expiresAt: addDays(new Date(), 1).toISOString(),
-};
-
 export {
   validLicense,
-  invalidLicense,
   saasLicense,
+  invalidLicense,
   commercialExpired,
   validNonCommercial,
-};
+} from '@camunda/mock-server-utils';
