+++ webapp/client/packages/mock-server-utils/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@camunda/mock-server-utils",
+	"version": "0.0.1",
+	"private": true,
+	"type": "module",
+	"license": "LicenseRef-Camunda-1.0",
+	"description": "Shared C8 API mock data for Camunda FE tests",
+	"sideEffects": false,
+	"files": [
+		"lib"
+	],
+	"exports": {
+		".": "./lib/index.ts",
+		"./mocks/currentUser": "./lib/mocks/currentUser.ts",
+		"./mocks/license": "./lib/mocks/license.ts"
+	},
+	"scripts": {
+		"typecheck": "tsc"
+	},
+	"devDependencies": {
+		"@camunda/camunda-api-zod-schemas": "^0.0.58",
+		"typescript": "5.9.3"
+	}
+}
