+++ webapp/client/packages/mock-server-utils/tsconfig.json
@@ -1,0 +1,22 @@
+{
+	"compilerOptions": {
+		"skipLibCheck": true,
+		"target": "ES2022",
+		"resolveJsonModule": true,
+		"moduleDetection": "force",
+		"isolatedModules": true,
+		"verbatimModuleSyntax": true,
+
+		"strict": true,
+		"noUncheckedIndexedAccess": true,
+		"noImplicitOverride": true,
+
+		"module": "preserve",
+		"moduleResolution": "Bundler",
+
+		"lib": ["ES2022"],
+
+		"noEmit": true
+	},
+	"include": ["lib"]
+}
